feat: Document everything
