+++ mnn-sys/src/lib.rs
@@ -11,6 +11,7 @@ mod sys {
     #![allow(non_upper_case_globals)]
     #![allow(non_camel_case_types)]
     #![allow(non_snake_case)]
+    #![allow(clippy::manual_c_str_literals)]
     include!(concat!(env!("OUT_DIR"), "/mnn_c.rs"));
 }
 pub use sys::*;
