+++ src/backend.rs
@@ -1,8 +1,14 @@
+//! The backend module contains the data types for the backend configuration
+
 use crate::prelude::*;
 use std::str::FromStr;
 
 use mnn_sys::*;
 
+/// BackendConfig is a struct that holds the configuration for the backend
+/// memory: [MemoryMode]
+/// power: [PowerMode]
+/// precision: [PrecisionMode]
 #[repr(transparent)]
 pub struct BackendConfig {
     pub(crate) inner: *mut MNNBackendConfig,
@@ -60,11 +66,15 @@ impl Default for BackendConfig {
     }
 }
 
+/// PowerModes depend on if the specific backend has support for it
 #[derive(Debug, Clone, Copy)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum PowerMode {
+    /// Low power mode
     Low,
+    /// Normal power mode
     Normal,
+    /// High power mode
     High,
 }
 
@@ -77,6 +87,7 @@ impl PowerMode {
         }
     }
 
+    /// Returns a string representation of the power mode
     pub fn to_str(self) -> &'static str {
         match self {
             Self::Low => "low",
@@ -139,11 +150,15 @@ impl FromStr for PrecisionMode {
     }
 }
 
+/// MemoryModes depend on if the specific backend has support for it
 #[derive(Debug, Clone, Copy)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum MemoryMode {
+    /// Low memory mode
     Low,
+    /// Normal memory mode
     Normal,
+    /// High memory mode
     High,
 }
 
@@ -156,6 +171,7 @@ impl MemoryMode {
         }
     }
 
+    /// Returns a string representation of the memory mode
     pub fn to_str(self) -> &'static str {
         match self {
             Self::Low => "low",
@@ -174,16 +190,21 @@ impl MemoryMode {
     }
 }
 
+/// PrecisionModes depend on if the specific backend has support for it
 #[derive(Debug, Clone, Copy)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum PrecisionMode {
+    /// Normal precision mode
     Normal = 0,
+    /// High precision mode
     High,
+    /// Low precision mode
     Low,
+    /// Low precision mode with BF16
     LowBf16,
 }
 impl PrecisionMode {
-    fn to_mnn_sys(self) -> mnn_sys::PrecisionMode {
+    pub(crate) fn to_mnn_sys(self) -> mnn_sys::PrecisionMode {
         match self {
             Self::LowBf16 => mnn_sys::PrecisionMode::Precision_Low_BF16,
             Self::Low => mnn_sys::PrecisionMode::Precision_Low,
@@ -192,6 +213,7 @@ impl PrecisionMode {
         }
     }
 
+    /// Returns a string representation of the precision mode
     pub fn to_str(self) -> &'static str {
         match self {
             Self::LowBf16 => "low_bf16",
@@ -231,11 +253,13 @@ impl BackendConfig {
         }
     }
 
+    /// Sets the [MemoryMode] for the backend
     pub fn with_memory_mode(mut self, mode: MemoryMode) -> Self {
         self.set_memory_mode(mode);
         self
     }
 
+    /// Gets the [MemoryMode] for the backend
     pub fn get_memory_mode(&self) -> MemoryMode {
         unsafe { MemoryMode::from_mnn_sys(mnn_sys::mnnbc_get_memory_mode(self.inner)) }
     }
@@ -247,11 +271,13 @@ impl BackendConfig {
         }
     }
 
+    /// Sets the [PowerMode] for the backend
     pub fn with_power_mode(mut self, mode: PowerMode) -> Self {
         self.set_power_mode(mode);
         self
     }
 
+    /// Gets the [PowerMode] for the backend
     pub fn get_power_mode(&self) -> PowerMode {
         unsafe { PowerMode::from_mnn_sys(mnn_sys::mnnbc_get_power_mode(self.inner)) }
     }
@@ -263,11 +289,13 @@ impl BackendConfig {
         }
     }
 
+    /// Sets the [PrecisionMode] for the backend
     pub fn with_precision_mode(mut self, mode: PrecisionMode) -> Self {
         self.set_precision_mode(mode);
         self
     }
 
+    /// Gets the [PrecisionMode] for the backend
     pub fn get_precision_mode(&self) -> PrecisionMode {
         unsafe { PrecisionMode::from_mnn_sys(mnn_sys::mnnbc_get_precision_mode(self.inner)) }
     }
@@ -280,6 +308,7 @@ impl BackendConfig {
         }
     }
 
+    /// Sets the flags for the backend
     pub fn with_flags(mut self, flags: usize) -> Self {
         self.set_flags(flags);
         self
@@ -294,6 +323,9 @@ impl BackendConfig {
         }
     }
 
+    /// # Safety
+    /// This just binds to the underlying unsafe api and should be used only if you know what you
+    /// are doing
     pub unsafe fn with_shared_context(mut self, shared_context: *mut libc::c_void) -> Self {
         self.set_shared_context(shared_context);
         self
